Refactor: Tag, Function js파일로 분리
