+++ js/tags.js
@@ -1,0 +1,8 @@
+const email = document.querySelector("#email");
+const password = document.querySelector("#password");
+const passwordVisible = document.querySelector(".toogle-eye");
+const emailError = document.createElement("div");
+const passwordError = document.createElement("div");
+const loginButton = document.querySelector(".signin-form");
+
+export { email, password, passwordVisible, emailError, passwordError, loginButton };
