+++ pages/signin.html
@@ -6,7 +6,7 @@
     <link rel="stylesheet" href="/css/reset.css" />
     <link rel="stylesheet" href="/css/variables.css" />
     <link rel="stylesheet" href="/css/signin.css" />
-    <script defer src="/js/signin.js"></script>
+    <script defer type="module" src="/js/signin.js"></script>
     <title>로그인 페이지</title>
   </head>
   <body>
